refactor: :recycle: move maps/flat_map into internals
